add AuthorController with service

+++ src/main/java/pl/sda/library/repository/AuthorRepository.java
@@ -3,6 +3,7 @@ package pl.sda.library.repository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import pl.sda.library.model.Author;
+import pl.sda.library.service.exceptions.ObjectDoesNotExistException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -37,7 +38,10 @@ public class AuthorRepository {
     }
 
     public void deleteAuthor(long id) {
-        authorDtoRepository.deleteById(id);
+        if(isExistById(id))
+            authorDtoRepository.deleteById(id);
+        else
+            throw new ObjectDoesNotExistException("Nie ma autora o takim ID: " + id);
     }
 
     public Author editAuthor(Author author) {
@@ -45,6 +49,13 @@ public class AuthorRepository {
         return savedAuthor.toModel();
     }
 
+    public boolean isExistById(Long id){
+        Optional<AuthorDto> byId = authorDtoRepository.findById(id);
+        if(byId.isPresent())
+            return true;
+        else
+            return false;
+    }
     public boolean isExist(String firstName, String lastName) {
         Optional<AuthorDto> authorDto = authorDtoRepository.findAuthorDtoByFirstNameAndLastName(firstName, lastName);
         if(authorDto.isPresent())
